+++ core-service/src/main/java/com/naumovets/market/core/entities/products/Product.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 @Entity
@@ -22,7 +23,7 @@ public class Product {
     private String title;
 
     @Column(name = "cost")
-    private int cost;
+    private BigDecimal cost;
 
     @ManyToOne
     @JoinColumn(name = "category_id")
