+++ core-service/src/main/resources/db/migration/V2__ordersData.sql
@@ -2,7 +2,7 @@ create table orders
 (
     id          bigserial primary key,
     username    varchar(255) not null,
-    total_price int not null,
+    total_price numeric(8, 2) not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -15,8 +15,8 @@ create table order_items
     product_id        bigint not null references products (id),
     order_id          bigint not null references orders (id),
     quantity          int    not null,
-    price_per_product int    not null,
-    price             int    not null,
+    price_per_product numeric(8, 2)    not null,
+    price             numeric(8, 2)    not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
